Create Token comparison.md
We aim to compare IOTA to other projects in a fair and well researched diagram / sheet. This should offer the viewer a quick and reliable comparison of IOTAS potential and features in relation to other projects. Please provide links to your sources for the collected metrics. so it can be verified that we dont claim wrong statistics.